requirements.txt
gym==0.23.1
numpy==1.23.1
pygame
scikit-image
tqdm
pyyaml
matplotlib
torch
torchvision
py_trees==2.1.6
opencv-python
pandas
seaborn
shapely
ephem
joblib
tensorboardX
cpprb
moviepy